requirements.txt
pandas
numpy
scikit-learn
xgboost
matplotlib
seaborn
flask
PyYAML
ensure==1.0.2
-e .